+++ src/main/kotlin/com/youquiz/user/global/config/SecurityConfiguration.kt
@@ -1,4 +1,4 @@
-package com.youquiz.template.global.config
+package com.youquiz.user.global.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
