+++ src/test/kotlin/com/youquiz/user/UserApplicationTest.kt
@@ -1,6 +1,6 @@
-package com.youquiz.template
+package com.youquiz.user
 
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class TemplateApplicationTest
+class UserApplicationTest
