+++ src/test/kotlin/com/youquiz/user/util/BaseControllerTest.kt
@@ -1,4 +1,4 @@
-package com.youquiz.template.util
+package com.youquiz.user.util
 
 import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.DescribeSpec
